added a bunch of reorganization stuff in the e2e tests to adapt them into a existing scenario (#3173)
