+++ Dockerfiles/e2e-tests/e2e-tests.Dockerfile
@@ -35,6 +35,13 @@ ENV E2E_TEST_DSC_MONITORING_NAMESPACE=opendatahub
 ENV E2E_TEST_CLEAN_UP_PREVIOUS_RESOURCES=false
 ENV E2E_TEST_DELETION_POLICY=never
 ENV E2E_TEST_FAIL_FAST_ON_ERROR=false
+ENV E2E_TEST_OPERATOR_CONTROLLER=true
+ENV E2E_TEST_OPERATOR_RESILIENCE=true
+ENV E2E_TEST_OPERATOR_V2TOV3UPGRADE=true
+ENV E2E_TEST_HARDWARE_PROFILE=true
+ENV E2E_TEST_WEBHOOK=true
+ENV E2E_TEST_COMPONENTS=true
+ENV E2E_TEST_SERVICES=true
 
 RUN apt-get update -y && apt-get upgrade -y && \
     curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl" && \
@@ -57,8 +64,4 @@ RUN mkdir -p results
 # run main go command
 CMD gotestsum --junitfile-project-name odh-operator-e2e \
 --junitfile results/xunit_report.xml --format testname --raw-command \
--- test2json -p e2e ./e2e-tests --test.v=test2json --test.parallel=8 \
---deletion-policy=$E2E_TEST_DELETION_POLICY --clean-up-previous-resources=$E2E_TEST_CLEAN_UP_PREVIOUS_RESOURCES \
---operator-namespace=$E2E_TEST_OPERATOR_NAMESPACE --applications-namespace=$E2E_TEST_APPLICATIONS_NAMESPACE \
---workbenches-namespace=$E2E_TEST_WORKBENCHES_NAMESPACE --dsc-monitoring-namespace=$E2E_TEST_DSC_MONITORING_NAMESPACE \
---fail-fast-on-error=$E2E_TEST_FAIL_FAST_ON_ERROR
+-- test2json -p e2e ./e2e-tests --test.v=test2json --test.parallel=8
